Extract outside path finder logic
Changes:
- Extract path finder logic from WalkOutside behind a new interface OutsidePathFinder
- Existing logic move to CollisionPathFinder as it is based on collistion detection
- Path logic held in a new dedicated walk package
- Add new method to LocalAreaUtil to hide the use of Line2D
- Create new unit test for OutsidePathFinder

Ref #1624
